+++ src/main/java/manzil/model/id/LikedPlaceId.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
 @Data
 public class LikedPlaceId implements Serializable
 {
-    private int user;
-    private int place;
+    private long user;
+    private long place;
 
 }
